requirements.txt
selenium==4.31.0
streamlit==1.32.0
python-dotenv==1.0.1